hover effect, cleanup and breaking out constant
