+++ src/components/municipalities/map/constants.ts
@@ -1,0 +1,7 @@
+export const MUNICIPALITY_MAP_COLORS = {
+  null: "var(--grey)",
+  start: "var(--pink-5)",
+  gradientMidLow: "var(--pink-4)",
+  gradientMidHigh: "var(--pink-3)",
+  end: "var(--blue-3)",
+} as const;
